telemetry_api/requirements.txt
Flask==3.0.2
influxdb-client[ciso]==1.41.0
paho-mqtt==2.0.0